--- a/src/bits/chunk_iterator/mod.rs
+++ b/src/bits/chunk_iterator/mod.rs
@@ -1,7 +1,6 @@
-mod chunk;
 mod merge;
 
-pub use chunk::BitChunk;
+pub use crate::types::BitChunk;
 use merge::merge_reversed;
 
 #[derive(Debug)]