@Dandandan 2021-04-22
Impressive! I found the crate with the same idea of using auto vectorization mentioned in the other PR:

https://docs.rs/slipstream/0.2.0/slipstream/

Maybe there is also some cool inspiration to find there.